refactor: extract prepareTestExecutionPlan() into reusable util
